:recycle: Refator field price

+++ src/main/domain/entities/products_entity.py
@@ -1,5 +1,5 @@
 from src.database.database import create_table, get_connect
-from peewee import Model, CharField, DecimalField
+from peewee import Model, CharField, IntegerField
 import uuid
 
 
@@ -7,7 +7,7 @@ class Product(Model):
 
     id = CharField(primary_key=True, default=str(uuid.uuid4()))
     name = CharField()
-    price = DecimalField()
+    price = IntegerField()
     description = CharField()
     tag = CharField()
